requirements.txt
fastapi
uvicorn
speechbrain
scipy
numpy
soundfile
passlib[bcrypt]
pyjwt
python-multipart
sqlalchemy
psycopg2-binary
dnspython
python-dotenv
redis
python-jose
librosa
pytest
pytest-cov
pydub
scipy
webrtcvad
noisereduce
httpx
pydub
webrtcvad
sounddevice